Move static data files into platform
