+++ src/js/common/containers/MHVApp.jsx
@@ -5,7 +5,7 @@ import { connect } from 'react-redux';
 import AcceptTermsPrompt from '../components/AcceptTermsPrompt';
 import AlertBox from '@department-of-veterans-affairs/jean-pants/AlertBox';
 import LoadingIndicator from '../components/LoadingIndicator';
-import { mhvAccessError } from '../utils/error-messages';
+import { mhvAccessError } from '../../../platform/static-data/error-messages';
 import {
   acceptTerms,
   createMHVAccount,
