+++ src/js/discharge-wizard/components/FormQuestions.jsx
@@ -7,7 +7,7 @@ import Scroll from 'react-scroll';
 import recordEvent from '../../../platform/monitoring/record-event';
 import ErrorableRadioButtons from '../../common/components/form-elements/ErrorableRadioButtons';
 import ErrorableSelect from '@department-of-veterans-affairs/jean-pants/ErrorableSelect';
-import { months } from '../../common/utils/options-for-select.js';
+import { months } from '../../../platform/static-data/options-for-select.js';
 import { focusElement } from '../../common/utils/helpers';
 import { questionLabels, prevApplicationYearCutoff, answerReview } from '../config';
 import { shouldShowQuestion } from '../utils';
