+++ src/js/post-911-gib-status/containers/Main.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 
 import LoadingIndicator from '../../common/components/LoadingIndicator';
-import { systemDownMessage, unableToFindRecordWarning } from '../../common/utils/error-messages';
+import { systemDownMessage, unableToFindRecordWarning } from '../../../platform/static-data/error-messages';
 
 import { getEnrollmentData } from '../actions/post-911-gib-status';
 // import { noChapter33BenefitsWarning } from '../utils/helpers.jsx';
