+++ src/js/rx/components/ErrorView.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { isEmpty, some, includes, intersection, concat } from 'lodash';
 
 import AlertBox from '@department-of-veterans-affairs/jean-pants/AlertBox';
-import { mhvAccessError } from '../../common/utils/error-messages';
+import { mhvAccessError } from '../../../platform/static-data/error-messages';
 import { errorCodes } from '../config';
 
 class ErrorView extends React.Component {
