+++ src/platform/startup/moment-setup.js
