+++ src/platform/static-data/labels.jsx
