+++ src/platform/static-data/options-for-select.js
